--- compiler/.ferret.json
@@ -1,18 +1,0 @@
-{
-  "compiler": {
-    "version": "0.1.0"
-  },
-  "cache": {
-    "path": ".ferret/modules"
-  },
-  "remote": {
-    "enabled": true,
-    "share": true
-  },
-  "dependencies": {
-      "modules": [
-          "code/remote/graphics",
-          "code/remote/audio"
-        ]
-  }
-}
